Copied necessary header files into imfitlib/include.
Updated setup.py so it checks to see if libimfit.a already exists in prebuilt/ and copies
it to libimfit/; if not, builds libimfit.a with SCons and then copies it to libimfit/.

Added flag for clang++ linker to suppress warnings (to avoid pages of "could not create compact unwind",
etc).

Added storage of version number in pyimfit/__init__.py, read by setup.py.
Updated version to 0.7.5.